RDM-10429:fixed sonar build issues
